@@ -9,19 +9,19 @@
     "tauri": "tauri"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4",
     "@tailwindcss/postcss": "^4.1.18",
     "@tailwindcss/vite": "^4.1.18",
     "@tauri-apps/cli": "^2.10.0",
     "@tsconfig/svelte": "^5.0.7",
     "autoprefixer": "^10.4.24",
     "postcss": "^8.5.6",
-    "svelte": "^4.2.7",
+    "svelte": "^5.51.3",
     "svelte-check": "^4.4.0",
     "svelte-preprocess": "^6.0.3",
     "tailwindcss": "^4.1.18",
     "typescript": "^5.9.3",
-    "vite": "^5.0.0"
+    "vite": "^5.4.4"
   },
   "dependencies": {
     "@tauri-apps/api": "^2.10.1",
