@@ -1,7 +1,13 @@
 import App from "./App.svelte";
+import { mount } from "svelte";
 
-const app = new App({
-  target: document.getElementById("app"),
+const target = document.getElementById("app");
+if (!target) {
+  throw new Error("App root element '#app' was not found");
+}
+
+const app = mount(App, {
+  target,
 });
 
 export default app;
